+++ lib/read/cardano-wallet-read.cabal
@@ -60,6 +60,7 @@ library
     Cardano.Read.Ledger.Tx.Metadata
     Cardano.Read.Ledger.Tx.Mint
     Cardano.Read.Ledger.Tx.ReferenceInputs
+    Cardano.Read.Ledger.Tx.ScriptValidity
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -92,7 +93,6 @@ library
     Cardano.Wallet.Read.Tx.Gen.Shelley
     Cardano.Wallet.Read.Tx.Gen.TxParameters
     Cardano.Wallet.Read.Tx.Inputs
-    Cardano.Wallet.Read.Tx.ScriptValidity
     Cardano.Wallet.Read.Tx.TxId
     Cardano.Wallet.Read.Tx.TxIn
     Cardano.Wallet.Read.Tx.Validity
